+++ .github/workflows/rust.yml
@@ -16,6 +16,6 @@ jobs:
     steps:
     - uses: actions/checkout@v3
     - name: Build
-      run: cargo build --verbose --manifest-path tensor_lib/Cargo.toml
+      run: cargo build --verbose
     - name: Run tests
-      run: cargo test --verbose --manifest-path tensor_lib/Cargo.toml
+      run: cargo test --verbose
