+++ app/Cargo.toml
@@ -6,4 +6,4 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-tensor_lib = { path = "../tensor_lib" }
+tensorlib = { path = "../tensorlib" }
