+++ tensorlib/src/advanced_ops/mod.rs
