+++ tensorlib/src/ops/equals.rs
