+++ tensorlib/src/ops/slice.rs
