+++ tensorlib/src/ops/slicelinear.rs
