+++ tensorlib/src/ops/subtract.rs
